feat!: adds ability to autostart QUnit tests
BREAKING CHANGE: To avoid losing test context, removes the ability to collect test results if the tests are already completed in the browser when the service is inserted. Instead, a recommendation is made to disable test auto-start in QUnit and allow the service to start the tests itself. If this is infeasible, adding a manual delay in the tests will also serve to give the service time to enter the browser context prior to test completion.